lux_housing_lookup_bot/requirements.txt
python-telegram-bot==21.9
gspread==6.1.4
google-auth==2.37.0
